fix(fetchrssproxy): error handling
